fix lint error by moving decode test function. caused by build tag.
