+++ v1/plugins/logs/eventBuffer.go
@@ -1,3 +1,7 @@
+// Copyright 2018 The OPA Authors.  All rights reserved.
+// Use of this source code is governed by an Apache2
+// license that can be found in the LICENSE file.
+
 package logs
 
 import (
